+++ snow-chat-im-backend/snow-im-api/src/main/java/com/stylesmile/chat/mapper/ChatGroupMemberMapper.java
@@ -22,5 +22,5 @@ public interface ChatGroupMemberMapper extends BaseMapper<ChatGroupMember> {
      * @param groupId 群组ID
      * @return 成员列表
      */
-    List<ChatGroupMember> getMembersByGroupId(@Param("groupId") Integer groupId);
+    List<ChatGroupMember> getMembersByGroupId(@Param("groupId") Long groupId);
 }
