+++ snow-chat-im-backend/snow-im-api/src/main/java/com/stylesmile/chat/service/ChatFriendRequestService.java
@@ -13,20 +13,20 @@ public interface ChatFriendRequestService extends BaseService<ChatFriendRequest>
     /**
      * 查询待处理的好友请求（接收到的）
      */
-    List<ChatFriendRequest> getPendingRequests(Integer toUserId);
+    List<ChatFriendRequest> getPendingRequests(Long toUserId);
 
     /**
      * 查询已发送的待处理请求
      */
-    List<ChatFriendRequest> getSentRequests(Integer fromUserId);
+    List<ChatFriendRequest> getSentRequests(Long fromUserId);
 
     /**
      * 发送好友请求
      */
-    void sendRequest(Integer fromUserId, Integer toUserId, String remark);
+    void sendRequest(Long fromUserId, Long toUserId, String remark);
 
     /**
      * 处理好友请求
      */
-    void handleRequest(Integer fromUserId, Integer toUserId, boolean accept);
+    void handleRequest(Long fromUserId, Long toUserId, boolean accept);
 }
