+++ snow-chat-im-backend/snow-im-api/src/main/java/com/stylesmile/chat/service/ChatFriendService.java
@@ -13,25 +13,25 @@ public interface ChatFriendService extends BaseService<ChatFriend> {
     /**
      * 查询用户的好友列表
      */
-    List<ChatFriend> getFriendsByUserId(Integer userId);
+    List<ChatFriend> getFriendsByUserId(Long userId);
 
     /**
      * 查询两个用户之间的好友关系
      */
-    ChatFriend getFriend(Integer userId, Integer friendId);
+    ChatFriend getFriend(Long userId, Long friendId);
 
     /**
      * 判断两个用户是否为好友
      */
-    boolean isFriend(Integer userId, Integer friendId);
+    boolean isFriend(Long userId, Long friendId);
 
     /**
      * 添加好友（幂等）
      */
-    void addFriend(Integer userId, Integer friendId);
+    void addFriend(Long userId, Long friendId);
 
     /**
      * 删除好友（双向删除）
      */
-    void removeFriend(Integer userId, Integer friendId);
+    void removeFriend(Long userId, Long friendId);
 }
